requirements.txt
langgraph==0.2.45
langchain-openai==0.2.9
langchain-core>=0.3.17,<0.4.0
pydantic==2.10.3
python-dotenv==1.0.1
streamlit==1.40.1